WIP WIP: Move pkg/a8n to enterprise/pkg/a8n
