+++ cmd/frontend/graphqlbackend/settings_subject.go
@@ -6,13 +6,12 @@ import (
 
 	graphql "github.com/graph-gophers/graphql-go"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
-	"github.com/sourcegraph/sourcegraph/pkg/a8n"
 	"github.com/sourcegraph/sourcegraph/pkg/api"
 	"github.com/sourcegraph/sourcegraph/pkg/jsonc"
 )
 
 func (r *schemaResolver) SettingsSubject(ctx context.Context, args *struct{ ID graphql.ID }) (*settingsSubject, error) {
-	return settingsSubjectByID(ctx, r.A8NStore, args.ID)
+	return settingsSubjectByID(ctx, r.A8NResolver, args.ID)
 }
 
 var errUnknownSettingsSubject = errors.New("unknown settings subject")
@@ -27,8 +26,8 @@ type settingsSubject struct {
 
 // settingsSubjectByID fetches the settings subject with the given ID. If the ID refers to a node
 // that is not a valid settings subject, an error is returned.
-func settingsSubjectByID(ctx context.Context, s *a8n.Store, id graphql.ID) (*settingsSubject, error) {
-	resolver, err := NodeByID(ctx, s, id)
+func settingsSubjectByID(ctx context.Context, a8n A8NResolver, id graphql.ID) (*settingsSubject, error) {
+	resolver, err := NodeByID(ctx, a8n, id)
 	if err != nil {
 		return nil, err
 	}
